Move version endpoint to app
